Service FinancialYearMasterService.groovy           RegionRegisterService.groovy           RoleMasterService.groovy           RuleMasterService.groovy           SeriesMasterService.groovy           TransactionTypeMaseterService.groovy

+++ phitb_entity/grails-app/domain/phitb_entity/SeriesMaster.groovy
@@ -9,7 +9,7 @@ class SeriesMaster {
     long returnId
     long orderId
     long status
-    int syncStatus
+    long syncStatus
     long entityTypeId
     long entityId
     long createdUser
